Creates landing page views
